refactor data sources and utils
